Use Tune-provided stoppers
