+++ tune_sklearn/tune_gridsearch.py
@@ -9,10 +9,9 @@ from sklearn.base import clone
 from sklearn.model_selection import ParameterGrid
 from ray import tune
 from tune_sklearn.list_searcher import ListSearcher
-from tune_sklearn.tune_utils import _IterTrialStopped
 from tune_sklearn.utils import (_check_param_grid_tune_grid_search,
                                 check_is_pipeline, check_error_warm_start,
-                                is_tune_grid_search)
+                                is_tune_grid_search, MaximumIterationStopper)
 from tune_sklearn.tune_basesearch import TuneBaseSearchCV
 from tune_sklearn._trainable import _Trainable
 from tune_sklearn._trainable import _PipelineTrainable
@@ -135,11 +134,8 @@ class TuneGridSearchCV(TuneBaseSearchCV):
             determined by 'Pipeline.warm_start' or 'Pipeline.partial_fit'
             capabilities, which are by default not supported by standard
             SKlearn. Defaults to True.
-        stop_on_plateau (bool|dict|TrialPlateauStopper): Stop trials early if
-            a plateau is reached. If False, disable. If True, enable with
-            default settings. If dict, overwrite default options of
-            `TrialPlateauStopper`. If `TrialPlateauStopper`, use the
-            passed object directly.
+        stopper (ray.tune.stopper.Stopper): Stopper objects passed to
+            ``tune.run()``.
         time_budget_s (int|float|datetime.timedelta): Global time budget in
             seconds after which all trials are stopped. Can also be a
             ``datetime.timedelta`` object.
@@ -161,7 +157,7 @@ class TuneGridSearchCV(TuneBaseSearchCV):
                  use_gpu=False,
                  loggers=None,
                  pipeline_auto_early_stop=True,
-                 stop_on_plateau=False,
+                 stopper=None,
                  time_budget_s=None,
                  sk_n_jobs=None):
         if sk_n_jobs is not None:
@@ -183,7 +179,7 @@ class TuneGridSearchCV(TuneBaseSearchCV):
             use_gpu=use_gpu,
             loggers=loggers,
             pipeline_auto_early_stop=pipeline_auto_early_stop,
-            stop_on_plateau=stop_on_plateau,
+            stopper=stopper,
             time_budget_s=time_budget_s)
 
         check_error_warm_start(self.early_stop_type, param_grid, estimator)
@@ -249,9 +245,9 @@ class TuneGridSearchCV(TuneBaseSearchCV):
         else:
             config["estimator_list"] = [self.estimator]
 
-        stopper = _IterTrialStopped(max_iter=self.max_iters)
-        if self.stop_on_plateau:
-            stopper = CombinedStopper(stopper, self.stop_on_plateau)
+        stopper = MaximumIterationStopper(max_iter=self.max_iters)
+        if self.stopper:
+            stopper = CombinedStopper(stopper, self.stopper)
 
         run_args = dict(
             scheduler=self.early_stopping,
